v2.9.1 - minor changes to support compiling on multiple platforms
git-svn-id: https://chyms.nrc.gc.ca/svn/raven/trunk@192 9a3899ca-7adb-4d27-9b38-7f4ebb3123b0

+++ src/RavenMain.cpp
@@ -140,7 +140,7 @@ int main(int argc, char* argv[])
 
         JulianConvert(t+Options.timestep,Options.julian_start_day,Options.julian_start_year,Options.calendar,tt);//increments time structure
         pModel->WriteMinorOutput(Options,tt);
-        pModel->WriteProgressOutput(Options,clock()-t1,step,ceil(Options.duration/Options.timestep));
+        pModel->WriteProgressOutput(Options,clock()-t1,step,(int)ceil(Options.duration/Options.timestep));
         //pModel->WriteProgressOutput(Options,clock()-t0,step+e*nsteps,nEnsembleMembers*nsteps); //TMP DEBUG - for ensemble support
 
         if(CheckForStopfile(step,tt)) { break; }
